requirements.txt
agno==2.0.4
numpy==2.3.2
openai==1.107.3
openpyxl==3.1.5
pandas==2.3.2
pgvector==0.4.1
pydantic==2.11.9
pypdf==6.0.0
python-dotenv==1.1.1
SQLAlchemy==2.0.43
psycopg[binary]